OS Project 2/requirements.txt
matplotlib>=3.5.0
pyyaml>=6.0
sqlite3>=2.6.0 